requirements.txt
opencv-python
torch
torchvision
Pillow
numpy
transformers
sentencepiece
ultralytics
streamlit-webrtc
